# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileRecord"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileRecord"
  /profiles/{id}:
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /maps:
    get:
      operationId: listMaps
      tags: [maps]
      summary: List all maps with pinpoints
      responses:
        "200":
          description: List of maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapRecord"
    post:
      operationId: createMap
      tags: [maps]
      summary: Create a map (moderator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapInput"
      responses:
        "201":
          description: Created map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapRecord"
  /maps/{id}:
    patch:
      operationId: updateMap
      tags: [maps]
      summary: Update a map (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapInput"
      responses:
        "200":
          description: Updated map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapRecord"
    delete:
      operationId: deleteMap
      tags: [maps]
      summary: Delete a map (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /maps/{mapId}/pinpoints:
    post:
      operationId: createPinpoint
      tags: [maps]
      summary: Add a pinpoint to a map (moderator)
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinpointInput"
      responses:
        "201":
          description: Created pinpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinpointRecord"
  /maps/{mapId}/pinpoints/{pinId}:
    patch:
      operationId: updatePinpoint
      tags: [maps]
      summary: Update a pinpoint (moderator)
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: integer
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinpointInput"
      responses:
        "200":
          description: Updated pinpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinpointRecord"
    delete:
      operationId: deletePinpoint
      tags: [maps]
      summary: Delete a pinpoint (moderator)
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: integer
        - name: pinId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProfileRecord:
      type: object
      required: [id, username, displayName, favoriteColor, bio, imageData, ageGroup, badges, traits, skin, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        username:
          type: string
        displayName:
          type: string
        favoriteColor:
          type: string
        bio:
          type: string
        imageData:
          type: string
        ageGroup:
          type: string
        badges:
          type: array
          items:
            type: string
        traits:
          type: array
          items:
            type: string
        skin:
          type: string
        banner:
          type: ["string", "null"]
        createdAt:
          type: string
    ProfileInput:
      type: object
      required: [username, displayName, favoriteColor, bio, imageData, ageGroup]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 30
        displayName:
          type: string
          minLength: 1
          maxLength: 30
        favoriteColor:
          type: string
        bio:
          type: string
          maxLength: 75
        imageData:
          type: string
        ageGroup:
          type: string
        banner:
          type: string
        traits:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        skin:
          type: string
        badges:
          type: array
          items:
            type: string
        banner:
          type: ["string", "null"]
        username:
          type: string
        displayName:
          type: string
        favoriteColor:
          type: string
        bio:
          type: string
        imageData:
          type: string
        ageGroup:
          type: string
        traits:
          type: array
          items:
            type: string
...
    MapRecord:
      type: object
      required: [id, name, imageData, createdAt, pinpoints]
      properties:
        id:
          type: integer
        name:
          type: string
        imageData:
          type: string
        createdAt:
          type: string
        pinpoints:
          type: array
          items:
            $ref: "#/components/schemas/PinpointRecord"
    MapInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        imageData:
          type: string
        accessCode:
          type: string
    PinpointRecord:
      type: object
      required: [id, mapId, type, name, description, imageData, xPercent, yPercent, createdAt, residents]
      properties:
        id:
          type: integer
        mapId:
          type: integer
        type:
          type: string
        name:
          type: string
        description:
          type: string
        imageData:
          type: string
        xPercent:
          type: number
        yPercent:
          type: number
        createdAt:
          type: string
        residents:
          type: array
          items:
            $ref: "#/components/schemas/ResidentRecord"
    PinpointInput:
      type: object
      required: [name]
      properties:
        type:
          type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        imageData:
          type: string
        xPercent:
          type: number
        yPercent:
          type: number
        accessCode:
          type: string
    ResidentRecord:
      type: object
      required: [userId, robloxUsername, displayName, description]
      properties:
        userId:
          type: integer
        robloxUsername:
          type: string
        displayName:
          type: string
        description:
          type: string
...