# lib/api-zod/src/generated/types/mapInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MapInput {
  /** @minLength 1 */
  name: string;
  imageData?: string;
  accessCode?: string;
}