# lib/api-zod/src/generated/types/mapRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PinpointRecord } from './pinpointRecord';

export interface MapRecord {
  id: number;
  name: string;
  imageData: string;
  createdAt: string;
  pinpoints: PinpointRecord[];
}