# lib/api-zod/src/generated/types/pinpointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PinpointInput {
  type?: string;
  /** @minLength 1 */
  name: string;
  description?: string;
  imageData?: string;
  xPercent?: number;
  yPercent?: number;
  accessCode?: string;
}