# lib/api-zod/src/generated/types/pinpointRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentRecord } from './residentRecord';

export interface PinpointRecord {
  id: number;
  mapId: number;
  type: string;
  name: string;
  description: string;
  imageData: string;
  xPercent: number;
  yPercent: number;
  createdAt: string;
  residents: ResidentRecord[];
}