# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  skin?: string;
  badges?: string[];
  /** @nullable */
  banner?: string | null;
  username?: string;
  displayName?: string;
  favoriteColor?: string;
  bio?: string;
  imageData?: string;
  ageGroup?: string;
  traits?: string[];
}